refactor(controller): isolate override validation rules
